chore(release): enforce release.branches at publish time
Address two regressions in the previous commit:

1. On main, the publish workflow was filtering to packages bumped in the
   latest release commit, which broke catch-up semantics: if a release ran
   and publish didn't, the next release+publish would skip the earlier
   package entirely. Restore the pre-existing behavior of publishing every
   eligible package on main; nx publish is a no-op for versions already on
   npm, so catch-up works again.

2. Gating only in tools/release.mjs left raw flows like `pnpm nx release
   version` and `pnpm nx release <v> --skip-publish` free to tag an
   ineligible package, after which the publish workflow would happily
   push it to latest. Move the hard gate into the publish workflow so the
   release.branches check is authoritative regardless of how versioning
   was performed.

Structural change: extract the eligibility helper into
tools/release-branches.mjs so tools/release.mjs (imports the function) and
.github/workflows/publish.yaml (invokes the CLI) share one implementation.

On prereleases, keep intersecting the eligible set with the commit-body
bumped list — that workaround is still required for the nx publish bugs
referenced in the workflow comments.
